reset and update to current stage

+++ fineuploader/endpoint-cors.php
@@ -54,15 +54,14 @@ $method = get_request_method();
 // send a DELETE request in a cross-origin environment in IE9 or older, it is not
 // possible to send a DELETE request.  So, we send a POST with the intended method,
 // DELETE, in a "_method" parameter.
-function get_request_method()
-{
+function get_request_method() {
     global $HTTP_RAW_POST_DATA;
 
     // This should only evaluate to true if the Content-Type is undefined
     // or unrecognized, such as when XDomainRequest has been used to
     // send the request.
-    if (isset($HTTP_RAW_POST_DATA)) {
-        parse_str($HTTP_RAW_POST_DATA, $_POST);
+    if(isset($HTTP_RAW_POST_DATA)) {
+    	parse_str($HTTP_RAW_POST_DATA, $_POST);
     }
 
     if (isset($_POST["_method"]) && $_POST["_method"] != null) {
@@ -72,29 +71,27 @@ function get_request_method()
     return $_SERVER["REQUEST_METHOD"];
 }
 
-function parseRequestHeaders()
-{
+
+function parseRequestHeaders() {
     $headers = array();
-    foreach ($_SERVER as $key => $value) {
+    foreach($_SERVER as $key => $value) {
         if (substr($key, 0, 5) <> 'HTTP_') {
             continue;
         }
-        $header           = str_replace(' ', '-', ucwords(str_replace('_', ' ', strtolower(substr($key, 5)))));
+        $header = str_replace(' ', '-', ucwords(str_replace('_', ' ', strtolower(substr($key, 5)))));
         $headers[$header] = $value;
     }
     return $headers;
 }
 
-function handleCorsRequest()
-{
+function handleCorsRequest() {
     header("Access-Control-Allow-Origin: *");
 }
 
 /*
  * handle pre-flighted requests. Needed for CORS operation
  */
-function handlePreflight()
-{
+function handlePreflight() {
     handleCorsRequest();
     header("Access-Control-Allow-Methods: POST, DELETE");
     header("Access-Control-Allow-Credentials: true");
@@ -103,7 +100,7 @@ function handlePreflight()
 
 // Determine whether we are dealing with a regular ol' XMLHttpRequest, or
 // an XDomainRequest
-$_HEADERS      = parseRequestHeaders();
+$_HEADERS = parseRequestHeaders();
 $iframeRequest = false;
 if (!isset($_HEADERS['X-Requested-With']) || $_HEADERS['X-Requested-With'] != "XMLHttpRequest") {
     $iframeRequest = true;
@@ -114,9 +111,12 @@ if (!isset($_HEADERS['X-Requested-With']) || $_HEADERS['X-Requested-With'] != "X
  */
 if ($method == "OPTIONS") {
     handlePreflight();
-} /*
+}
+
+/*
  * handle a DELETE request or a POST with a _method of DELETE.
- */ elseif ($method == "DELETE") {
+ */
+else if ($method == "DELETE") {
     handleCorsRequest();
 
     $result = $uploader->handleDelete("files");
@@ -125,13 +125,14 @@ if ($method == "OPTIONS") {
     // return some JSON along with self-executing javascript (iframe.ss.response)
     // that will parse the JSON and pass it along to Fine Uploader via
     // window.postMessage
-    if ($iframeRequest === true) {
+    if ($iframeRequest == true) {
         header("Content-Type: text/html");
-        echo json_encode($result) . "<script src='http://10.0.2.2/jquery.fineuploader-4.1.1/iframe.xss.response-4.1.1.js'></script>";
+        echo json_encode($result)."<script src='http://10.0.2.2/jquery.fineuploader-4.1.1/iframe.xss.response-4.1.1.js'></script>";
     } else {
         echo json_encode($result);
     }
-} elseif ($method == "POST") {
+}
+else if ($method == "POST") {
     handleCorsRequest();
     header("Content-Type: text/plain");
 
@@ -139,7 +140,8 @@ if ($method == "OPTIONS") {
     // For example: /myserver/handlers/endpoint.php?done
     if (isset($_GET["done"])) {
         $result = $uploader->combineChunks("files");
-    } // Handles upload requests
+    }
+    // Handles upload requests
     else {
         // Call handleUpload() with the name of the folder, relative to PHP's getcwd()
         $result = $uploader->handleUpload("files");
@@ -151,13 +153,16 @@ if ($method == "OPTIONS") {
         // return some JSON along with self-executing javascript (iframe.ss.response)
         // that will parse the JSON and pass it along to Fine Uploader via
         // window.postMessage
-        if ($iframeRequest === true) {
+        if ($iframeRequest == true) {
             header("Content-Type: text/html");
-            echo json_encode($result) . "<script src='http://{{SERVER_URL}}/{{FINE_UPLOADER_FOLDER}}/iframe.xss.response.js'></script>";
+            echo json_encode($result)."<script src='http://{{SERVER_URL}}/{{FINE_UPLOADER_FOLDER}}/iframe.xss.response.js'></script>";
         } else {
             echo json_encode($result);
         }
     }
-} else {
+}
+else {
     header("HTTP/1.0 405 Method Not Allowed");
 }
+
+?>
